pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tradercat"
version = "0.1.0"
description = "A multi-symbol, multi-strategy trading bot"
authors = [{ name = "AimeSoleil", email = "jeaimesoleil@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openbb==4.3.5",
    "pydantic>=2.11.7,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "aiohttp>=3.12.15,<4.0.0",
    "apscheduler>=3.10.4",
    "pytz>=2024.1",
    "tabulate>=0.9.0"
]

[project.optional-dependencies]
dev = [
    "black==25.1.0",
    "isort==5.13.2",
    "mypy==1.10.0",
    "pytest==8.2.1",
    "pytest-asyncio==0.23.6",
    "flake8==7.0.0"
]

[project.scripts]
tradercat = "trade_bot.main:main"